fix(test): extract EVAL_FIXTURES to standalone module so eval-tutor.ts script does not crash
Closes the InitSuite crash @thierry hit when running
`npx tsx scripts/eval-tutor.ts` with a real OPENROUTER_API_KEY:

  TypeError: Cannot read properties of undefined (reading 'config')
  at initSuite (.../@vitest/runner/dist/chunk-artifact.js:1848:23)
  at <anonymous> (src/test/ai/evalSuite.test.ts:219:1)

Root cause: the EVAL_FIXTURES corpus and its types were co-located in
src/test/ai/evalSuite.test.ts which also declares `describe`/`it` Vitest
hooks at module top-level. Importing that file from a standalone tsx
process pulled the Vitest globals which only resolve inside the runner,
hence the InitSuite crash on import.

Fix: extract `EvalCategory`, `EvalFixture`, and `EVAL_FIXTURES` into a
dedicated `src/test/ai/eval-fixtures.ts` (pure data + types, zero
runtime side effects). The two consumers now import from there:

- src/test/ai/evalSuite.test.ts — keeps the `describe`/`it` block
  (CI gate against structural prompt regression)
- scripts/eval-tutor.ts — standalone Haiku run (eval suite b),
  loadable outside the Vitest runner

Verification:
- `npx vitest run src/test/ai/` → 286/286 AI tests passing (no
  regression)
- `npx vitest run src/test/ai/evalSuite.test.ts` → 18/18 (corpus +
  cue assertions intact)
- `npx tsc --noEmit` → clean
- `npx tsx scripts/eval-tutor.ts` (no env) → expected error message
  `❌ Missing OPENROUTER_API_KEY` instead of the InitSuite crash —
  module loads cleanly, script ready to run with a real key

Refs:
- Bug report: @thierry runtime test, 10 May 2026 PM
- ADR-008 § "Eval suite hybrid" — gate ship still requires (b) manual
  run before each prompt bump
- Empirical runtime verdict on v1.1.0 production:
  · F1 compound: PARTIAL (3 bullets correctement, mais 3 questions
    finales — comportement conforme au prompt actuel, à reformuler en
    V1.1.1 si UX cible veut < 3 questions)
  · F2 over-explanation: PASS NET (refus drift kernel/syscalls)
  · F3 repeated-hint: pending eval (b) multi-turn
  · F4 satisfaction signal: PASS NET (cliché éliminé)
